dev_requirements.txt
-r test_requirements.txt
wheel
django-compass2
bpython

